cleanup crate structure
